create a separate code file for each filter class
